fix: rename folder to quantum_teleportation
